requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
psycopg[binary]==3.2.10
alembic==1.13.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.3.0
httpx==0.28.1
pytest==8.4.2
pytest-cov==5.0.0
